created utilities folder, refactored transactions
