+++ app/src/components/ViewAccounts.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom'
-import FormatCurrency from './FormatCurrency';
+import FormatCurrency from '../app_utilities/FormatCurrency';
 
 const ViewAccounts = () => {
     const [accounts, setAccounts] = useState(null);
