+++ app/src/pages/accounts/Accounts.js
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom'
 import { MdOutlineEdit, MdAddCircleOutline, MdDeleteOutline } from "react-icons/md";
-import FormatCurrency from '../../components/FormatCurrency';
+import FormatCurrency from '../../app_utilities/FormatCurrency';
+import { handleError } from '../../app_utilities/HandleError';
 
 const GetAccounts = () => {
     const navigate = useNavigate();
@@ -13,7 +14,7 @@ const GetAccounts = () => {
         fetchAccounts();
     }, []);
 
-    const fetchAccounts = async() => {
+    const fetchAccounts = async () => {
         try {
             const token = localStorage.getItem('token');
             const response = await axios.get('http://127.0.0.1:8000/peppermint/account/my_accounts', {
@@ -23,10 +24,7 @@ const GetAccounts = () => {
             });
             setAccounts(response.data);
         } catch (error) {
-            console.error('Error retrieving accounts.', error);
-            if (error.response.status === 401) {
-                navigate('/login');
-            }
+            handleError(error, navigate);
         }
     };
