+++ app/src/pages/transactions/Transactions.js
@@ -2,11 +2,13 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom'
 import { MdOutlineEdit, MdAddCircleOutline, MdDeleteOutline } from "react-icons/md";
-import FormatCurrency from '../../components/FormatCurrency';
-import FormatDate from '../../components/FormatDate';
+import FormatCurrency from '../../app_utilities/FormatCurrency';
+import FormatDate from '../../app_utilities/FormatDate';
+import { handleError } from '../../app_utilities/HandleError';
 
 const GetAllTransactions = () => {
     const navigate = useNavigate();
+    const getToken = () => localStorage.getItem('token');
     const [transactions, setTransactions] = useState([]);
     const [accounts, setAccounts] = useState(null);
 
@@ -22,18 +24,14 @@ const GetAllTransactions = () => {
 
     const fetchAccounts = async() => {
         try {
-            const token = localStorage.getItem('token');
             const response = await axios.get('http://127.0.0.1:8000/peppermint/account/my_accounts', {
                 headers: {
-                    Authorization: `Bearer ${token}`
+                    Authorization: `Bearer ${getToken()}`
                 }
             });
             setAccounts(response.data);
         } catch (error) {
-            console.error('Error retrieving accounts.', error);
-            if (error.response.status === 401) {
-                navigate('/login');
-            }
+            handleError(error, navigate);
         }
     };
 
@@ -41,38 +39,35 @@ const GetAllTransactions = () => {
 
     const fetchAccountTransactions =  async (account_id) => {
         try {
-            const token = localStorage.getItem('token');
             const response = await axios.get(`http://127.0.0.1:8000/peppermint/account/${account_id}/transactions`, {
                     headers: {
-                         Authorization: `Bearer ${token}`
+                         Authorization: `Bearer ${getToken()}`
                     }
                 });
             return response.data || [];
             
         } catch (error) {
-            console.error('Error retrieving accounts.', error);
-            if (error.response.status === 401) {
-                navigate('/login');
+            handleError(error, navigate);
             return [];
-            }
         }
     };
 
+
     const fetchAllTransactions = async () => {
         const accountIds = getAccountIds();
-        const allTransactions = [];
+        try {
+            const transactionPromises = accountIds.map(fetchAccountTransactions);
+            const allTransactions = (await Promise.all(transactionPromises)).flat();
 
-        for (const id of accountIds) {
-            const transactions = await fetchAccountTransactions(id);
-            allTransactions.push(...transactions);
-        }
-
-        if (allTransactions.length === 0) {
-            alert('No Transactions available at this time');
-            setTransactions([]);
-        } else {
-            const sortedTransactions = allTransactions.sort((a,b) => new Date(b.transaction_date) - new Date(a.transaction_date));
-            setTransactions(sortedTransactions);
+            if (allTransactions.length === 0) {
+                alert('No Transactions available at this time');
+                setTransactions([]);
+            } else {
+                const sortedTransactions = allTransactions.sort((a,b) => new Date(b.transaction_date) - new Date(a.transaction_date));
+                setTransactions(sortedTransactions);
+            }
+        } catch (error) {
+            handleError(error, navigate);
         }
     };
 
@@ -82,19 +77,15 @@ const GetAllTransactions = () => {
 
     const handleDeleteTransaction = async(account_id, id) => {
         try {
-            const token = localStorage.getItem('token');
             await axios.delete(`http://127.0.0.1:8000/peppermint/${account_id}/${id}`, {
                 headers: {
-                    Authorization: `Bearer ${token}`
+                    Authorization: `Bearer ${getToken()}`
                 }
             });
-            fetchAllTransactions();
+            await fetchAllTransactions();
             alert('Transaction deleted!')
         } catch (error) {
-            console.error('Error retrieving transactions', error);
-            if (error.response.status === 401) {
-                navigate('/login');
-            }
+            handleError(error, navigate);
         }
     };
